fix: added some logs
Current problem, there is only one active session in selenium grid.
Maybe the nodes urls should be a direct internal link to the specific
machine, not the app (because grid thinks all the machines is one node),
or simply expose a different port for each machine. If that doesn't
work, try just updating the "SE_NODE_MAX_SESSIONS" and see if it spreads
the sessions through all the different machines.