feat(SeleniumGridAutoscaler): create without starting machine, to get the identifier, update the node url environment variable to the specific machine url
Specific machine url is used so that seleniumgrid knows which machine to register, to prevent it thinking each machine in fly nodes app is one single node. The url of the machine, accessible to other applications in the organization through private networking, includes the machine identifier. There is no way to specify machine identifier when creating, fly generates the identifier.